+++ _import/tycho-execution/tests/protocol_integration_tests.rs
@@ -42,6 +42,7 @@ fn test_single_encoding_strategy_ekubo() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFrom);
@@ -95,6 +96,7 @@ fn test_single_encoding_strategy_maverick() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFrom);
@@ -160,6 +162,7 @@ fn test_single_encoding_strategy_usv4_eth_in() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFromPermit2);
 
@@ -227,6 +230,7 @@ fn test_single_encoding_strategy_usv4_eth_out() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFromPermit2);
@@ -302,6 +306,7 @@ fn test_single_encoding_strategy_usv4_grouped_swap() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let swap_eth_pepe = Swap {
@@ -316,6 +321,7 @@ fn test_single_encoding_strategy_usv4_grouped_swap() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFromPermit2);
 
@@ -425,6 +431,7 @@ fn test_single_encoding_strategy_curve() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFrom);
@@ -493,6 +500,7 @@ fn test_single_encoding_strategy_curve_st_eth() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFrom);
@@ -547,6 +555,7 @@ fn test_single_encoding_strategy_balancer_v3() {
         split: 0f64,
         user_data: None,
         protocol_state: None,
+        estimated_amount_in: None,
     };
 
     let encoder = get_tycho_router_encoder(UserTransferType::TransferFrom);